[frontend] fix case value type for advanced conf
