+++ opencti-platform/opencti-front/src/private/components/settings/sso_definitions/OpenIDConfig.tsx
@@ -110,85 +110,6 @@ const OpenIDConfig = ({ updateField }: OpenIDConfigProps) => {
           </>
         )}
       </FieldArray>
-      <FieldArray name="advancedConfigurations">
-        {({ push, remove, form }) => (
-          <>
-            <div style={{ display: 'flex', alignItems: 'center', marginTop: 20 }}>
-              <Typography variant="h2">{t_i18n('Add more fields')}</Typography>
-              <IconButton
-                size="default"
-                color="secondary"
-                // size="large"
-                style={{ marginBottom: 8 }}
-                onClick={() => push({ key: '', value: '', type: 'String' })}
-              >
-                <Add fontSize="small" color="primary" />
-              </IconButton>
-            </div>
-            {form.values.advancedConfigurations
-              && form.values.advancedConfigurations.map(
-                (
-                  conf: { key: string; value: string; type: string },
-                  index: number,
-                ) => (
-                  <div
-                    key={index}
-                    style={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      justifyContent: 'space-around',
-                      marginBottom: 8,
-                    }}
-                  >
-                    <Field
-                      component={TextField}
-                      variant="standard"
-                      onSubmit={() => updateField('advancedConfigurations', form.values.advancedConfigurations)}
-                      name={`advancedConfigurations[${index}].key`}
-                      label={t_i18n('Key (in passport)')}
-                      containerstyle={{ width: '20%' }}
-                    />
-                    <Field
-                      component={TextField}
-                      variant="standard"
-                      onSubmit={() => updateField('advancedConfigurations', form.values.advancedConfigurations)}
-                      name={`advancedConfigurations[${index}].value`}
-                      label={t_i18n('Value (in IDP)')}
-                      containerstyle={{ width: '20%' }}
-                    />
-                    <Field
-                      component={SelectField}
-                      variant="standard"
-                      onSubmit={() => updateField('advancedConfigurations', form.values.advancedConfigurations)}
-                      name={`advancedConfigurations[${index}].type`}
-                      label={t_i18n('Field type')}
-                      containerstyle={{ width: '20%' }}
-                    >
-                      <MenuItem value="Boolean">Boolean</MenuItem>
-                      <MenuItem value="Integer">Integer</MenuItem>
-                      <MenuItem value="String">String</MenuItem>
-                      <MenuItem value="Array">Array</MenuItem>
-                    </Field>
-                    <IconButton
-                      color="primary"
-                      aria-label={t_i18n('Delete')}
-                      style={{ marginTop: 10 }}
-                      size="default"
-                      onClick={() => {
-                        remove(index);
-                        const advancedConfigurations = [...form.values.advancedConfigurations];
-                        advancedConfigurations.splice(index, 1);
-                        updateField('advancedConfigurations', advancedConfigurations);
-                      }} // Delete
-                    >
-                      <Delete fontSize="small" />
-                    </IconButton>
-                  </div>
-                ),
-              )}
-          </>
-        )}
-      </FieldArray>
     </>
   );
 };
